handle TODO1 todos
